+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-
+- [Core] Update `<ListRow>` & `<List>` styles for supporting nested list better. (#257)
 
 ## [4.1.0]