+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+### Changed
 - [Core] Update `<ListRow>` & `<List>` styles for supporting nested list better. (#257)
+- [Core] Fix `rowProps` used to be replaced in mixin `formRow`.(#258)
 
 ## [4.1.0]